MNT: move single_gen and ensure_generator to utils
