+++ bluesky/utils.py
@@ -65,6 +65,56 @@ class PlanHalt(GeneratorExit):
     pass
 
 
+PLAN_TYPES = (types.GeneratorType,)
+try:
+    from types import CoroutineType
+except ImportError:
+    # < py35
+    pass
+else:
+    PLAN_TYPES = PLAN_TYPES + (CoroutineType, )
+    del CoroutineType
+
+
+def ensure_generator(plan):
+    """
+    Ensure that the input is a generator.
+
+    Parameters
+    ----------
+    plan : iterable or iterator
+
+    Returns
+    -------
+    gen : coroutine
+    """
+    gen = iter(plan)  # no-op on generators; needed for classes
+    if not isinstance(gen, PLAN_TYPES):
+        # If plan does not support .send, we must wrap it in a generator.
+        gen = (msg for msg in gen)
+
+    return gen
+
+
+def single_gen(msg):
+    '''Turn a single message into a plan
+
+    If ``lambda x: yield x`` were valid Python, this would be equivalent.
+    In Python 3.6 or 3.7 we might get lambda generators.
+
+    Parameters
+    ----------
+    msg : Msg
+        a single message
+
+    Yields
+    ------
+    msg : Msg
+        the input message
+    '''
+    yield msg
+
+
 class SignalHandler:
     def __init__(self, sig, log=None):
         self.sig = sig
